Extract interface from ValueSnapshotter
